fs: hardcode sizeof tmpbuf
As clearly pointer out by compiler: ``` argument to 'sizeof' in 'snprintf' call is the same expression as the destination ```

Here we are trying to find sizeof a char array
But because it's not initialised with some size
it's basically returning a pointer to the 0th element
and then that's basically sizeof char.

But the pointer is initialised with 128 bits of memory
anyway on declaration so hardcode the size here.

Signed-off-by: CannedShroud <hypersync112@gmail.com>